refactor: share job source helpers
